requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
blinker==1.4
boto3==1.24.1
botocore==1.27.1
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cycler==0.11.0
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.0.12
Flask==2.0.1
Flask-Mail==0.9.1
flask-paginate==0.8.1
Flask-SQLAlchemy==2.5.1
fonttools==4.28.5
greenlet==1.1.0
gTTS==2.2.3
gunicorn==20.1.0
idna==2.10
image==1.5.33
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Js2Py==0.71
kiwisolver==1.3.2
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
matplotlib==3.5.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
mysql==0.0.3
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
paytmchecksum==1.7.0
paywix==1.4.0
Pillow==8.3.1
pipwin==0.5.1
platformdirs==2.2.0
portalocker==2.3.2
psycopg2==2.9.3
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycryptodome==3.14.1
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypi==2.1
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
python-dateutil==2.8.2
PyYAML==5.4.1
razorpay==1.3.0
render==1.0.0
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stringcase==1.2.0
template==0.7.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.7.0
Werkzeug==2.0.1
whitenoise==6.1.0